requirements.txt
streamlit
google-generativeai
pinecone
pypdf
python-dotenv
tqdm
